+++ website/src/components/confirmation-dialog.css.ts
@@ -1,12 +1,5 @@
 import { style } from "@vanilla-extract/css"
-import {
-  hspace,
-  mediaQueries,
-  radii,
-  space,
-  thickness,
-  vspace,
-} from "src/style/constants"
+import { mediaQueries, space, vspace } from "src/style/constants"
 
 export const buttons = style({
   display: "flex",
@@ -22,25 +15,4 @@ export const buttons = style({
       marginTop: 0,
     },
   },
-})
-
-export const cancelButton = style({
-  width: "8.625rem",
-  height: "1.9375rem",
-  cursor: "pointer",
-  border: `${thickness.thin} solid #5c3b37`,
-  borderRadius: radii.large,
-  backgroundColor: "#ffffff",
-  font: `normal 400 0.9375rem/1.2 'Charis SIL'`,
-})
-
-export const confirmButton = style({
-  width: "8.625rem",
-  height: "1.9375rem",
-  cursor: "pointer",
-  border: `${thickness.thin} solid #5c3b37`,
-  borderRadius: radii.large,
-  backgroundColor: "#415373",
-  color: "white",
-  font: `normal 400 0.9375rem/1.2 'Charis SIL'`,
 })
